Made frustum near value less than far
Abandoned VKguide convention to use a frustum near value greater than the far value to optimize z values distribution in the NDC in order to increase accuracy and prevent z-fighting.
Now the near value is expected to be less than the far value because it is more intuitive and because it follows the convention of the majority of tutorials and implementations.